make order fixed

+++ cheshire.mk
@@ -91,8 +91,7 @@ $(CHS_ROOT)/hw/newusb_regs/newusb_regs.hjson:
 
 # NewUSB registers
 $(CHS_ROOT)/hw/newusb_regs/newusb_reg_pkg.sv $(CHS_ROOT)/hw/newusb_regs/newusb_reg_top.sv: $(CHS_ROOT)/hw/newusb_regs/newusb_regs.hjson
-	$(REGTOOL) -r $< --outdir $(dir $@)
-	rm -f $(CHS_ROOT)/hw/newusb_regs/newusb_regs.hjson
+	$(REGTOOL) -r $< --outdir $(dir $@) && rm -f $(CHS_ROOT)/hw/newusb_regs/newusb_regs.hjson
 
 # CLINT
 CLINTCORES ?= 1
